+++ lib/karafka/pro/loader.rb
@@ -74,8 +74,8 @@ module Karafka
         #   that need to have some special configuration stuff injected into config, etc
         def features
           [
-            Encryption,
             Cleaner,
+            Encryption,
             RecurringTasks,
             ScheduledMessages
           ]