+++ lib/karafka/pro/routing/features/adaptive_iterator/contracts/topic.rb
@@ -32,12 +32,39 @@ module Karafka
                 required(:active) { |val| [true, false].include?(val) }
                 required(:safety_margin) { |val| val.is_a?(Integer) && val.positive? && val < 100 }
                 required(:adaptive_margin) { |val| [true, false].include?(val) }
-                required(:mark_after_yielding) { |val| [true, false].include?(val) }
                 required(:clean_after_yielding) { |val| [true, false].include?(val) }
 
                 required(:marking_method) do |val|
                   %i[mark_as_consumed mark_as_consumed!].include?(val)
                 end
+              end
+
+              # Since adaptive iterator uses `#seek` and can break processing in the middle, we
+              # cannot use it with virtual partitions that can process data in a distributed
+              # manner
+              virtual do |data, errors|
+                next unless errors.empty?
+
+                adaptive_iterator = data[:adaptive_iterator]
+                virtual_partitions = data[:virtual_partitions]
+
+                next unless adaptive_iterator[:active]
+                next unless virtual_partitions[:active]
+
+                [[%i[adaptive_iterator], :with_virtual_partitions]]
+              end
+
+              # There is no point of using the adaptive iterator with LRJ because of how LRJ works
+              virtual do |data, errors|
+                next unless errors.empty?
+
+                adaptive_iterator = data[:adaptive_iterator]
+                long_running_jobs = data[:long_running_job]
+
+                next unless adaptive_iterator[:active]
+                next unless long_running_jobs[:active]
+
+                [[%i[adaptive_iterator], :with_long_running_job]]
               end
             end
           end